feat: use named exports BREAKING CHANGE: changes export signature
