+++ tests/specs/paths.spec.ts
@@ -1,7 +1,6 @@
 import path from 'path';
 import { testSuite, expect } from 'manten';
-import getTsconfig from '../../dist/index.js';
-import { createPathsMatcher } from '../../dist/paths/index.js';
+import { getTsconfig, createPathsMatcher } from '../../dist/index.js';
 import { createFixture, tsconfigJson } from '../utils/create-fixture';
 
 export default testSuite(({ describe }) => {
